requirements.txt
joblib
pandas==2.2.2
pyyaml==6.0.2
pytest
mlflow
matplotlib
scikit-learn
numpy